README! Mustard Fried
Added some initial info that I wrote to my friend about the idea for the project. I edited it a little bit to work in the repo README format a little better. It's not stuctured very well, but that's alright. It's a nice example of letting your brain spill out ideas into words, just to get them out there and into the world!