Burger Builder
Was watching a ThePrimeagen video and I realized that the `BurgerBuilder` pattern would actually work out really nice for this project! It was funny how it was also about burgers in the example, hehe.

https://www.youtube.com/watch?v=ZfG8BSTX0Lw

Trying out on setting up a similar pattern here to the one used in the video.