Condiment Types Shift
Thinking about moving condiments away from being dedicated types, to instead being a distinct condiment type that's present as a property on the thing that you are ordering it on.

I can't decide how I would like this part to work at the type level though, I don't like having a `boolean | string` union type for that here, it feels a bit code-stinky. One idea I just thought of is using an enum-ish thing to correlate the condiment intensity instead using a level-based system with numbers instead. That feels a bit weird too though. Would 0 be none, 1 be extra light, and 6 be double extra? Or should 1 be regular, 0 be none, and then what? Hmm. That's what I liked about the `string | null` pattern, but then the redundant case there would be to have `'topping'` as the value for when the topping is present. Hmm hmm. That's what I liked about the original string interpolation version, it was more consistently using strings.

Gonna do more ideas testing! I think the next thing I realized is that I want to provide a JSON API that will fully show all of the data about your order. Then I could make an INO serializer of some sort that turns that into dedicated JS objects again. idk, something like that.

Oh yeah, and the full order would be an array of all the items in the order, and each thing could be it's own object with properties on it. I think if I try to think about it like this way instead, it will help me with trying to design an API to get the data into that format.