Burger Builder Interfacing
Got the burger builder setup working initially! Now you can modify the object with function chaining calls, which makes it really easy to dynamically modify your burger description.

I realized that this setup can possibly fit extremely well alongside using interfaces to define the different burger primitives, and using the base burger builder to construct each burger from any input. This allows you to also change the entry on the fly and it will correct itself before you "seal" the burger. I think I will probably try and do it this way for the other burger primitives too, so that you can just specify a given burger by it's condiments, meat, cheese, and buns presence, and it will in turn get you the correct burger primitive based on those criteria.

This is just like how are menu system works, and how we would probably take your order if you didn't specify all of the main burger primitives up front. On the menus at least, you have to specify which burger you would like to have before you change any of the condiments on it, so I may have to make the burger builder class private, since that's how the order taking process works too.

I think I will use error handling and console warnings to mirror how putting the burger in works when you change things like making a Double-Single into a Double-Double by adding Extra Cheese, it would show a `console.warn()` message in the console for that, haha. And if you try to have 6 patties on one burger, it would instead throw an error. This further makes me question how the error handling should work for those method chaining parts. Should this be Promise-based so the customer can handle the errors? As that's what we would do if they tried to order 6 patties. We would bubble the error from the order screen to them and provide them with the goal of deciding what to do with that error information, haha again.

But maybe it should just be a type-error kind of thing, and all of those kinds of errors should just be warnings. Maybe us ordertakers are simply the linters in the IDE for the programmer (the customer), and we are just letting them know about the type errors along the way. Hmmmm. Because say if someone wanted to order a Cheeseburger but they said Hamburger, ... with cheese, then I would press the Extra Cheese button. I wouldn't throw a //// Realized I changed my mind while writing that. I will show a `console.warn()` for this in the console (adding cheese to a Hamburger), but my original sentence is still right. I wouldn't warn a customer that a Hamburger with cheese is a Cheeseburger, I would simply read the end result as a Cheeseburger to them. There isn't anything to handle there error-wise, you simply would want to inform them of the 'console warning' per se.

*One edit to make:

For the number of patties part, I think I wouldn't throw for that, but rather just log a `console.error()` instead. This shows as a full error instead of just a warning //// Actually, that means there's no programmatic way to catch it! This is a sticky situation, going to have to work some things out here like that.