API Demos
Crazy mind rant with HoldYourWaffle on different ways to possibly represent the data using various programming methodologies.

A new possible idea I had for this is to make a receipt scanner, which will generate an order object for you, and you could make an order-editing GUI based on that JSON-ish representation of the order. So the idea is to find a way to represent an order using plain JSON.

A side idea with this is creating a way to make a text format which represents an order, kind of like SNBT. It would likely instead be a rendered image of the order which follows the same 'spec' that receipts do, and that could be the intermediate format of things.

So it could be an app which helps you share your order with friends, by scanning your receipt after your order. Then you could AirDrop (or something) your receipt to a friend, and they could edit their copy of that same order.